+++ src/index.ts
@@ -1,14 +1,7 @@
-import { Client, Options } from 'tmi.js';
-import fs from 'fs';
-import path from 'path';
-
 import OBSController from './OBSController';
-import Command from './Command';
-import SoundCommand from './SoundCommand';
-import SayCommand from './SayCommand';
-import Slippi from './Slippi.js';
-
-import * as twitchPermissionsTokens from '../twitchPermissionsTokens.json';
+import Slippi from './Slippi';
+import { client } from './tmiClient';
+import { sayInChat, findCommandInMessage } from './utils';
 
 if (process.argv.includes('--help')) {
   console.log('Usage: node index.js [--slippi]');
@@ -21,99 +14,18 @@ process.on('unhandledRejection', (reason, p) => {
   // application specific logging, throwing an error, or other logic here
 });
 
-const botName = 'DeceBot';
-const tmiOptions:Options = {
-  options: {
-    debug: false,
-  },
-  connection: {
-    reconnect: true,
-  },
-  identity: {
-    username: botName,
-    password: twitchPermissionsTokens[botName],
-  },
-  channels: ['itsdece'],
-};
-
-const commands = [
-  // Text to speech command:
-  new SayCommand('!Say', (message) => {
-    sayInChat(message);
-  }),
-
-  // DeceBot text commands:
-  new Command('!Discord', () => {
-    sayInChat('Join the discord here: https://discord.gg/65jUQ8G');
-  }),
-  new Command('!Github', () => {
-    sayInChat(
-      'Wanna see what my insides look like? https://github.com/micantre/DeceBot'
-    );
-  }),
-  new Command('!CupheadWars', () => {
-    sayInChat(
-      'The great Cuphead Wars of 2020 is a competition between me and a couple of my coworkers to see who can beat cuphead the fastest. Others in the competition are https://www.twitch.tv/kishkishftw and https://www.twitch.tv/faultymuse'
-    );
-  }),
-  new Command('!NewSoundboard', () => {
-    sayInChat(
-      'The subscriber sound effect commands are now shared in one collective subscriber soundboard.'
-    );
-  }),
-
-  // Sound commands are autimatically added in the code below
-];
-
-// Automatically add Sound Commands code from: https://stackoverflow.com/questions/32511789/looping-through-files-in-a-folder-node-js
-(async () => {
-  const soundBoardPath =
-    'C:\\Users\\Michael\\Desktop\\StreamStuff\\Media\\SoundEffects\\SoundBoard';
-  const folderPaths = [
-    soundBoardPath + '\\General',
-    soundBoardPath + '\\SubOnly',
-  ];
-  try {
-    for (const folderPath of folderPaths) {
-      const files = await fs.promises.readdir(folderPath);
-      const isSubOnly = folderPath.includes('SubOnly');
-      const commandsInFolder = [];
-      for (const file of files) {
-        if (file.includes('.txt')) {
-          continue;
-        }
-        const fullPath = path.join(folderPath, file);
-        const command = `!${file.split('.')[0]}`;
-        commands.push(
-          new SoundCommand(command, fullPath, sayInChat, isSubOnly)
-        );
-        commandsInFolder.push(command);
-      }
-      fs.writeFile(
-        folderPath + '\\Commands.txt',
-        commandsInFolder.join('\n'),
-        (error) => {}
-      );
-    }
-  } catch (e) {
-    console.error('Exception thrown while generating sound commands:', e);
-  } finally {
-    console.log('Loaded all sound commands in %s', soundBoardPath);
-  }
-})();
-
-const client = new (Client as unknown as {new(opts:Options):Client})(tmiOptions);
-client.connect();
-
-client.on('connected', function (a, p) {
-  sayInChat('SUP WORLD!');
-});
-
 const obsController = new OBSController();
 if (process.argv.includes('--slippi')) {
   console.log('String slippi task...');
   const slippi = new Slippi('D:\\Games\\Dolphin\\Slippi\\Games', obsController);
 }
+
+// These handlers could arguably be move to tmiClient.ts
+// We'll leave them here for now
+client.on('connected', function (a, p) {
+  sayInChat('SUP WORLD!');
+});
+
 
 client.on('chat', function (channel, userstate, message, self) {
   const command = findCommandInMessage(message);
@@ -122,16 +34,4 @@ client.on('chat', function (channel, userstate, message, self) {
   }
 });
 
-function findCommandInMessage(message: string) {
-  for (const command of commands) {
-    if (message.toLowerCase().startsWith(command.command.toLowerCase())) {
-      return command;
-    }
-  }
-  return null;
-}
-
-function sayInChat(message: string) {
-  client.say('itsdece', message);
-  // client.action("itsdece", message);
-}
+client.connect();
